ASF cleanup
I forgot some things, it seems my file was wrong, I used ffmpeg to encode to "wma" which copied the artwork of the source file as a video with 1000 fps, it really was a video. I just tried a true "wma" file and added cover art with MP3Tag and it works correctly, seeking and all.