Setting for handling interlacing in progressive mode
Only for MKV files with the interlaced flag. "Keep" displays the frames as is, "TV" switches to interlaced output to allow the TV to bob it, or "Blur" which does vertical blurring to hide the combing artifacts.

The 'enhanced resolution' setting must be enabled or set to auto.

This option will not work if you have 240p or 576p enabled.

This is the only way to support deinterlacing without affecting performance.