Add ffmpeg recipe for BPL30
Baseline Profile 3.0 is a generic recipe that the Netflix Wii app used for playing content. It's also used to provide comp. to old android devices.

These are not magic settings that will look good on everything. But is a good base to work with.

The other recipe is a CRF version that may seem better, but could hinder performance due to vbv limits not being as strict as in 2-pass.