Add some low mem subfonts
Avenir is good for subfont.ttf, though it the character size might be a bit small.

Consolas is a monospaced font, but not too wide, it good for both subfont.ttf and monospace.ttf

Rodin is a decent subfont.

VCR OSD is the best monospaced font for accurate EIA-608 closed captions. Rips made with recent ffmpeg versions can make use of this font to accurately portray positions. It also looks super cool.