A wildcard can now be used for url entries that use the search type
This is useful for streamlink. If you add an entry to onlinemedia.xml that's like this: "http://192.168.0.7:8080" the 7 can change for various reasons but the port doesn't have to. If you add type="search" to the entry you can use the keyboard to finish the url, but this would mean you'd have to write the port each time!

Here's where the wildcard comes into play, if you write the address like this: "http://192.168.0.*:8080" the text you write with the keyboard will replace the asterisk so now you'd only need to write the last octet of the IP of the computer running streamlink.