Screensaver banners no longer care about filenames
Once a folder is configured, it will scan for image extensions and use the total amount to build a shuffled list of entries.

After all images are shown, a new shuffle is done.

It's fine to put folders in the banner dir, but I wouldn't recommend putting other files here, the image detection isn't fully done yet, and WiiMC may crash trying to load other files.